+++ GPE/GPE_scalar_field_1d2c.py
@@ -13,7 +13,7 @@ class GPE_scalar_field_1d2c:
         
         
         
-        self.psi = np.zeros(self.my_shape,dtype=np.complex64)
+        self.psi = np.zeros(self.my_shape,dtype=np.complex128)
         self.psi = 1.0*ini_psi
 
         self.mass_ini = np.sum(np.abs(ini_psi)**2) 
@@ -25,7 +25,7 @@ class GPE_scalar_field_1d2c:
         
         self.f = 1.0*self.psi
         
-        self.im_K = np.zeros(self.K_shape,dtype=np.complex64)
+        self.im_K = np.zeros(self.K_shape,dtype=np.complex128)
         self.ex_K = np.zeros_like(self.im_K)
